# ode-wrapper/src/main/java/com/trihydro/odewrapper/docs/swagger-ui-master/dist/swagger.json
...
swagger: '2.0'
info:
  description: API Endpoints to create and manage CV TIMs from WYDOT Applications.
  version: 1.0.0
  title: WYDOT Connected Vehicle REST API
  termsOfService: http://swagger.io/terms/
  contact:
    email: kperry@trihydro.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: cvodepp01:7777
...
schemes:
- http
paths:
...
  "/rw-tim":
    get:
      tags:
      - Road Construction
      summary: Returns all road construction active TIMs
      operationId: getRoadConstructionTims
      produces:
      - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            "$ref": "#/definitions/GetResponse"
    post:
      tags:
      - Road Construction
      summary: Add road construction TIMs
      operationId: addRoadConstructionTim
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Road Construction fields
        required: true
        schema:
          "$ref": "#/definitions/RoadConstructionObj"
      responses:
        '405':
          description: Invalid input
    put:
      tags:
      - Road Construction
      summary: Update Road Construction TIMs
      operationId: updateRoadConstructionTim
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: Road Construction fields
        required: true
        schema:
          "$ref": "#/definitions/RoadConstructionObj"
      responses:
        '405':
          description: Invalid input
  "/rw-tim/itis-codes/{clientId}":
    get:
      tags:
      - Road Construction
      summary: Find road construction tim by ID with ITIS Codes
      operationId: getRWICById
      produces:
      - application/json
      parameters:
      - name: clientId
        in: path
        description: ID of Road Consttuction TIM
        required: true
        type: string
      responses:
        '200':
          description: successful operation
          schema:
            "$ref": "#/definitions/GetResponse"
        '400':
          description: Invalid ID supplied
        '404':
          description: RoadConstruction not found
  "/rw-tim/{clientId}":
    get:
      tags:
      - Road Construction
      summary: Find road construction tim by ID
      operationId: getRWById
      produces:
      - application/json
      parameters:
      - name: clientId
        in: path
        description: ID of Road Consttuction TIM
        required: true
        type: string
      responses:
        '200':
          description: successful operation
          schema:
            "$ref": "#/definitions/GetResponse"
        '400':
          description: Invalid ID supplied
        '404':
          description: RoadConstruction not found
    delete:
      tags:
      - Road Construction
      summary: Delete road construction tim by ID
      operationId: deleteRW
      produces:
      - application/json
      parameters:
      - name: clientId
        in: path
        description: ID of the Road Construction that needs to be deleted
        required: true
        type: string
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: Road Construction not found
definitions:
...
  RoadConstructionObj:
    properties:
      timRwList:
        type: array
        items:
          "$ref": "#/definitions/RoadConstruction"
  RoadConstruction:
    type: object
    required:
    - fromRm
    - toRm
    - highway
    - pk
    - clientId
    - direction
    - surface
    - buffers
    - startTs
    properties:
      fromRm:
        type: number
        example: 350
      toRm:
        type: number
        example: 360
      highway:
        type: string
        example: I-80
      pk:
        type: string
        example: '15917'
      clientId:
        type: string
        example: '15917'
      direction:
        type: string
        example: westbound
      surface:
        type: string
        example: P
      startTs:
        type: string
        example: '2018-04-16T19:30:05.000Z'
      buffers:
        type: array
        items:
          "$ref": "#/definitions/Buffer"
  Buffer:
    type: object
    required:
    - distance
    - action
    properties:
      distance:
        type: number
        example: 1
      action:
        type: string
        example: inform
...
  GetResponse:
    type: object
    properties:
      activeTimId:
        type: integer
        format: int32
        example: 884
      timId:
        type: integer
        format: int32
        example: 1642778
      milepostStart:
        type: integer
        format: int32
        example: 355
      milepostStop:
        type: integer
        format: int32
        example: 360
      direction:
        type: string
        example: westbound
      startDateTime:
        type: string
        example: '2018-04-16 19:30:05'
      endDateTime:
        type: string
        example: '2018-04-17 19:30:05'
      route:
        type: string
        example: '80'
      clientId:
        type: string
        example: IN1234
      satRecordId:
        type: string
        example: 01ACDA1A
      pk:
        type: string
        example: 123
...